requirements.txt
#pip install -r requirements.txt

django==4.2.7
my-sql-client==2.2.0
mysql-connector-python==8.2.0
django-environ==0.11.2